+++ Modules/SharePointDsc/DSCResources/MSFT_SPServiceAppSecurity/MSFT_SPServiceAppSecurity.psm1
@@ -125,7 +125,7 @@ function Set-TargetResource
                                 
                 if ($CurrentValues.Members.Username -contains $desiredMember.Username) {
                     # The user already has a permission, check that it is correct
-                    if (($CurrentValues.Members | Where-Object { $_.Username -eq $desiredMember.Username } | Select -First 1).AccessLevel -ne $desiredMember.AccessLevel) {
+                    if (($CurrentValues.Members | Where-Object { $_.Username -eq $desiredMember.Username } | Select-Object -First 1).AccessLevel -ne $desiredMember.AccessLevel) {
                         # Current permission for user doesn't match what it should, remove them and re-add them        
                         Revoke-SPObjectSecurity -Identity $security -Principal $claim
                         Grant-SPObjectSecurity -Identity $security -Principal $claim -Rights $desiredMember.AccessLevel
@@ -162,7 +162,7 @@ function Set-TargetResource
                 }
                 if ($CurrentValues.Members.Username -contains $desiredMember.Username) {
                     # The user already has a permission, check that it is correct
-                    if (($CurrentValues.Members | Where-Object { $_.Username -eq $desiredMember.Username } | Select -First 1).AccessLevel -ne $desiredMember.AccessLevel) {
+                    if (($CurrentValues.Members | Where-Object { $_.Username -eq $desiredMember.Username } | Select-Object -First 1).AccessLevel -ne $desiredMember.AccessLevel) {
                         # Current permission for user doesn't match what it should, remove them and re-add them
                         Revoke-SPObjectSecurity -Identity $security -Principal $claim
                         Grant-SPObjectSecurity -Identity $security -Principal $claim -Rights $desiredMember.AccessLevel
@@ -238,7 +238,7 @@ function Test-TargetResource
         if ($differences -eq $null) {
             Write-Verbose "Security list matches - checking that permissions match on each object"
             foreach($currentMember in $CurrentValues.Members) {
-                if ($currentMember.AccessLevel -ne ($Members | Where-Object { $_.Username -eq $currentMember.Username } | Select -First 1).AccessLevel) {
+                if ($currentMember.AccessLevel -ne ($Members | Where-Object { $_.Username -eq $currentMember.Username } | Select-Object -First 1).AccessLevel) {
                     Write-Verbose "$($currentMember.Username) has incorrect permission level. Test failed."
                     return $false
                 }
@@ -259,7 +259,7 @@ function Test-TargetResource
                 $result = $false
             } else {
                 Write-Verbose "$($member.Username) already has access. Checking permission..."
-                if ($member.AccessLevel -ne ($CurrentValues.Members | Where-Object { $_.Username -eq $member.Username } | Select -First 1).AccessLevel) {
+                if ($member.AccessLevel -ne ($CurrentValues.Members | Where-Object { $_.Username -eq $member.Username } | Select-Object -First 1).AccessLevel) {
                     Write-Verbose "$($member.Username) has incorrect permission level. Test failed."
                     return $false
                 }